+++ docs/Whats-New.md
@@ -172,6 +172,7 @@ New:
 - Implement `CellSpread` and `PercentAtMax` (by ZivDero)
 - Implement `ScorchChance`, `CraterChance` and `CellAnimChance` (by ZivDero)
 - Implement `Explosive` for animations (by ZivDero, Rampastring)
+- Implement `InfantryMultiplier`, `VehicleMultiplier`, `AircraftMultiplier`, `BuildingMultiplier`, `TerrainMultiplier` (by ZivDero)
 
 
 Vanilla fixes: